+++ CHANGELOG.md
@@ -30,8 +30,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Transform `mjml` files (email's template files) to html and plaintext 
 - Add a participants tab, listing all users connected to the chat
 - Add an authentificatin pop-up in the chat
-
->>>>>>> 💄(frontend) add svg icons needed for the feature
+- Add a default config file for the chat
 
 ### Changed